Revert "build OLM operator registry in systemtests (#4356)"
This reverts commit d4d9209ecdcb7b9d31aa5cdc7d7ce4c9d0dce70e.

+++ systemtests/src/test/java/io/enmasse/systemtest/isolated/upgrade/UpgradeTest.java
@@ -19,6 +19,7 @@ import io.enmasse.systemtest.bases.isolated.ITestIsolatedStandard;
 import io.enmasse.systemtest.condition.OpenShift;
 import io.enmasse.systemtest.condition.OpenShiftVersion;
 import io.enmasse.systemtest.executor.Exec;
+import io.enmasse.systemtest.executor.ExecutionResultData;
 import io.enmasse.systemtest.logs.CustomLogger;
 import io.enmasse.systemtest.logs.GlobalLogCollector;
 import io.enmasse.systemtest.messagingclients.AbstractClient;
@@ -43,8 +44,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.slf4j.Logger;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
 
 import java.io.IOException;
@@ -52,6 +51,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
@@ -77,7 +77,6 @@ class UpgradeTest extends TestBase implements ITestIsolatedStandard {
     private String infraNamespace;
     private EnmasseInstallType type;
     private OLMInstallationType olmType;
-    private EnmasseOperatorManager operatorManager = EnmasseOperatorManager.getInstance();
 
     @BeforeAll
     void prepareUpgradeEnv() throws Exception {
@@ -93,7 +92,7 @@ class UpgradeTest extends TestBase implements ITestIsolatedStandard {
         }
         boolean waitForNamespace = true;
         if (this.type.equals(EnmasseInstallType.BUNDLE)) {
-            assertTrue(operatorManager.clean());
+            assertTrue(EnmasseOperatorManager.getInstance().clean());
         } else if (this.type.equals(EnmasseInstallType.OLM)) {
             if (EnmasseOperatorManager.getInstance().isEnmasseOlmDeployed()) {
                 for (var addrSpace : kubernetes.getAddressSpaceClient(infraNamespace).list().getItems()) {
@@ -106,12 +105,12 @@ class UpgradeTest extends TestBase implements ITestIsolatedStandard {
                     }
                 }
             }
-            assertTrue(operatorManager.removeOlm());
+            assertTrue(EnmasseOperatorManager.getInstance().removeOlm());
             if (olmType != null && olmType == OLMInstallationType.DEFAULT) {
                 waitForNamespace = false;
             }
         } else {
-            operatorManager.deleteEnmasseAnsible();
+            EnmasseOperatorManager.getInstance().deleteEnmasseAnsible();
         }
         if (waitForNamespace) {
             TestUtils.waitForNamespaceDeleted(kubernetes, infraNamespace);
@@ -141,7 +140,7 @@ class UpgradeTest extends TestBase implements ITestIsolatedStandard {
     void testUpgradeOLMSpecific(String version, String templates) throws Exception {
         this.type = EnmasseInstallType.OLM;
         this.olmType = OLMInstallationType.SPECIFIC;
-        this.infraNamespace = operatorManager.getNamespaceByOlmInstallationType(olmType);
+        this.infraNamespace = EnmasseOperatorManager.getInstance().getNamespaceByOlmInstallationType(olmType);
         doTestUpgrade(templates, version);
     }
 
@@ -151,7 +150,7 @@ class UpgradeTest extends TestBase implements ITestIsolatedStandard {
     void testUpgradeOLMDefault(String version, String templates) throws Exception {
         this.type = EnmasseInstallType.OLM;
         this.olmType = OLMInstallationType.DEFAULT;
-        this.infraNamespace = operatorManager.getNamespaceByOlmInstallationType(olmType);
+        this.infraNamespace = EnmasseOperatorManager.getInstance().getNamespaceByOlmInstallationType(olmType);
         doTestUpgrade(templates, version);
     }
 
@@ -434,9 +433,37 @@ class UpgradeTest extends TestBase implements ITestIsolatedStandard {
     }
 
     private void installEnmasseOLM(Path templateDir, String version) throws Exception {
-        String manifestsImage = getManifestsImage(templateDir, version);
+
+        if (olmType == OLMInstallationType.SPECIFIC) {
+            kubernetes.createNamespace(infraNamespace, Collections.singletonMap("allowed", "true"));
+        }
+
+        String catalogSourceName;
+        String catalogNamespace;
+        if (version.startsWith("0.30")) {
+            catalogSourceName = "community-operators";
+            catalogNamespace = "openshift-marketplace";
+        } else {
+            catalogSourceName = "enmasse-source";
+            catalogNamespace = infraNamespace;
+
+            String customRegistryImageToUse = buildPushCustomOperatorRegistry(catalogNamespace, templateDir, version);
+
+            deployCatalogSource(catalogSourceName, catalogNamespace, customRegistryImageToUse);
+        }
+
+        if (olmType == OLMInstallationType.SPECIFIC) {
+            Path operatorGroupFile = Files.createTempFile("operatorgroup", ".yaml");
+            String operatorGroup = Files.readString(Paths.get("custom-operator-registry", "operator-group.yaml"));
+            Files.writeString(operatorGroupFile, operatorGroup.replaceAll("\\$\\{OPERATOR_NAMESPACE}", infraNamespace));
+            KubeCMDClient.applyFromFile(infraNamespace, operatorGroupFile);
+        }
+
         String csvName = getCsvName(templateDir, version);
-        operatorManager.olm().install(olmType, manifestsImage, csvName);
+
+        applySubscription(infraNamespace, catalogSourceName, catalogNamespace, csvName);
+
+        TestUtils.waitForPodReady("enmasse-operator", infraNamespace);
 
         Thread.sleep(30_000);
 
@@ -444,43 +471,91 @@ class UpgradeTest extends TestBase implements ITestIsolatedStandard {
         KubeCMDClient.applyFromFile(infraNamespace, Paths.get(templateDir.toString(), "install", "components", "example-authservices", "standard-authservice.yaml"));
 
         Thread.sleep(60_000);
-        operatorManager.waitUntilOperatorReadyOlm(olmType);
+        EnmasseOperatorManager.getInstance().waitUntilOperatorReadyOlm(olmType);
         Thread.sleep(30_000);
+    }
+
+    private void applySubscription(String installationNamespace, String catalogSourceName, String catalogNamespace, String csvName) throws IOException {
+        Path subscriptionFile = Files.createTempFile("subscription", ".yaml");
+        String subscription = Files.readString(Paths.get("custom-operator-registry", "subscription.yaml"));
+        Files.writeString(subscriptionFile,
+                subscription
+                    .replaceAll("\\$\\{OPERATOR_NAMESPACE}", installationNamespace)
+                    .replaceAll("\\$\\{CATALOG_SOURCE_NAME}", catalogSourceName)
+                    .replaceAll("\\$\\{CATALOG_NAMESPACE}", catalogNamespace)
+                    .replaceAll("\\$\\{CSV}", csvName));
+        KubeCMDClient.applyFromFile(installationNamespace, subscriptionFile);
+    }
+
+    private void deployCatalogSource(String catalogSourceName, String catalogNamespace, String customRegistryImageToUse) throws IOException {
+        Path catalogSourceFile = Files.createTempFile("catalogsource", ".yaml");
+        String catalogSource = Files.readString(Paths.get("custom-operator-registry", "catalog-source.yaml"));
+        Files.writeString(catalogSourceFile,
+                catalogSource
+                    .replaceAll("\\$\\{CATALOG_SOURCE_NAME}", catalogSourceName)
+                    .replaceAll("\\$\\{OPERATOR_NAMESPACE}", catalogNamespace)
+                    .replaceAll("\\$\\{REGISTRY_IMAGE}", customRegistryImageToUse));
+        KubeCMDClient.applyFromFile(catalogNamespace, catalogSourceFile);
     }
 
     private void upgradeEnmasseOLM(Path upgradeTemplates, String previousVersion) throws Exception {
         String newVersion = getVersionFromTemplateDir(upgradeTemplates);
-        String manifestsImage = getManifestsImage(upgradeTemplates, newVersion);
-        //update manifests image
-        operatorManager.olm().buildPushCustomOperatorRegistry(infraNamespace, manifestsImage);
+        String customRegistryImageToUse = buildPushCustomOperatorRegistry(infraNamespace, upgradeTemplates, newVersion);
 
-        //delete catalog pod to force update
-        kubernetes.deletePod(infraNamespace, Map.of("olm.catalogSource", "enmasse-source"));
-
-        String csvName = getCsvName(upgradeTemplates, newVersion);
         String catalogSourceName = "enmasse-source";
         String catalogNamespace = infraNamespace;
+        if (previousVersion.startsWith("0.30")) {
+            deployCatalogSource(catalogSourceName, catalogNamespace, customRegistryImageToUse);
+        } else {
+            kubernetes.deletePod(infraNamespace, Map.of("olm.catalogSource", "enmasse-source"));
+        }
+
+        String csvName = getCsvName(upgradeTemplates, newVersion);
+
         //update subscription to point to new catalog and to use latest csv
-        operatorManager.olm().applySubscription(infraNamespace, catalogSourceName, catalogNamespace, csvName);
+        applySubscription(infraNamespace, catalogSourceName, catalogNamespace, csvName);
 
         //should update
         Thread.sleep(300_000);
         checkImagesUpdated(getVersionFromTemplateDir(upgradeTemplates));
     }
 
-    private String getManifestsImage(Path templateDir, String version) throws IOException, JsonProcessingException, JsonMappingException {
-        String manifestsImage;
+    private String buildPushCustomOperatorRegistry(String namespace, Path templateDir, String version) throws Exception {
+        String customRegistryImageToPush = environment.getClusterExternalImageRegistry()+"/"+namespace+"/systemtests-operator-registry:latest";
+        String customRegistryImageToUse = environment.getClusterInternalImageRegistry()+"/"+namespace+"/systemtests-operator-registry:latest";
+
+        String olmManifestsImage;
         if (version.equals("1.3")) {
             String exampleCatalogSource = Files.readString(Paths.get(templateDir.toString(), "install", "components", "enmasse-operator", "050-Deployment-enmasse-operator.yaml"));
+            log.info(exampleCatalogSource);
+
+            log.info("Enmasse operator deployment found : {}", exampleCatalogSource!=null && !exampleCatalogSource.isEmpty());
             var yaml = new YAMLMapper().readTree(exampleCatalogSource);
             var spec = yaml.get("spec").get("template").get("spec");
-            manifestsImage = spec.get("containers").get(0).get("image").asText();
+            log.info("Spec fields");
+            spec.fieldNames().forEachRemaining(log::info);
+            olmManifestsImage = spec.get("containers").get(0).get("image").asText();
         } else {
             String exampleCatalogSource = Files.readString(Paths.get(templateDir.toString(), "install", "components", "example-olm", "catalog-source.yaml"));
             var tree = new YAMLMapper().readTree(exampleCatalogSource);
-            manifestsImage = tree.get("spec").get("image").asText();
+            olmManifestsImage = tree.get("spec").get("image").asText();
         }
-        return manifestsImage;
+
+        olmManifestsImage = olmManifestsImage.replace(environment.getClusterInternalImageRegistry(), environment.getClusterExternalImageRegistry());
+
+        int retries = 5;
+        ExecutionResultData results = null;
+        while (retries > 0) {
+            results = Exec.execute(Arrays.asList("make", "-C", "custom-operator-registry", "FROM="+olmManifestsImage, "TAG="+customRegistryImageToPush), true);
+            if(results.getRetCode()) {
+                return customRegistryImageToUse;
+            }
+            Thread.sleep(1000);
+            retries--;
+        }
+        assertTrue(results != null && results.getRetCode(), "custom operator registry image build failed ");
+
+        return customRegistryImageToUse;
     }
 
     private String getCsvName(Path templateDir, String version) throws Exception {
